+++ java/com.sap.sailing.server/src/com/sap/sailing/server/impl/RacingEventServiceMXBeanImpl.java
@@ -21,7 +21,7 @@ public class RacingEventServiceMXBeanImpl implements RacingEventServiceMXBean {
     }
 
     @Override
-    public int getNumberOfTrackedRacesToRestore() {
+    public long getNumberOfTrackedRacesToRestore() {
         return getRacingEventService().getNumberOfTrackedRacesToRestore();
     }
